pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
shortuuid = "^1.0.1"
firebase-admin = "^5.0.3"
firebase = "^3.0.1"
weasyprint = "^53.3"
xhtml2pdf = "^0.2.5"
pandas = "^1.3.4"
cryptography = "^35.0.0"
[tool.poetry.dev-dependencies]
